+++ Controllers/Especialidades/EspecialidadesDeleteController.cs
@@ -18,6 +18,11 @@ namespace simulacro2.Controllers.Especialidades
         [Route("api/especialidades/delete/{id}")]
         public IActionResult Delete(int id)
         {
+            //buscamos primero y verificar que este en la base de datos
+            var result = _repository.GetById(id);
+            if (result == null)
+                return NotFound($"No se encontro la especialidad con ese id {id}");
+            
             _repository.Delete(id);
             Response.Headers.Add("X-Message", "Especialidad eliminada correctamente");
             return StatusCode(204); // Devolver un 204 No Content para "delete"
